Voxel.java: changed float variables to double type. This change lead to many modifications in other classes. AbstractVoxelAnalysis.java: added new abstract function computeTransmittance(Voxel) VoxelAnalysis.java: updated so that it implements new abstract function computeTransmittance. Kept mode 2 and 3 for memory but not used anymore NewVoxelAnalysis.java: new class that reflects Gregoire improvements in the estimation of the transmittance. RSPVoxelisationService.java: updated to run with both VoxelAnalysis and NewVoxelAnalysis transmittance algorithms.

+++ AMAPVox/AMAPVoxCore/src/main/java/fr/amap/lidar/amapvox/chart/VoxelsToChart.java
@@ -572,7 +572,7 @@ public class VoxelsToChart {
             Voxel voxel = iterator.next();
             
             //float pad = voxel.calculatePAD(maxPAD);
-            float pad = voxel.PadBVTotal;
+            float pad = (float) voxel.PadBVTotal;
             
             if(pad > maxPAD){
                 pad = maxPAD;
